Create 3.3 Persistence Script
Purpose
The objective of this lab is to illustrate the concept of persistence on a Linux system through a user-level approach. Persistence ensures the consistent execution of a specific action or script, often following a system reboot.  In this lab, you will explore a basic form of user-level persistence on Linux using a user's shell configuration file.

Instructions

Understand the Target:
- Identify the user account for which you want to establish persistence.

Create a Script:
- Develop a script, in the language of your choice, with the necessary commands or actions you want to persist. Make sure it is executable (chmod +x YOUR SCRIPT).

Configure Automatic Execution:
- Find out how to make your script execute automatically. Research methods in Linux that allow a script to run upon login or periodically. Choose one method that suits your preference.

Testing:
- Log out and log back in, or open a new terminal.
- Observe if your script executes as expected.